final - added commands to add/remove words, check word list

+++ bot.py
@@ -3,6 +3,7 @@ import random
 from discord.ext import commands
 
 client = commands.Bot(command_prefix="!") #prefix for commands like !remove
+blacklist = ["frick", "dang", "darn", "damn", "heck", "crap", "shizzle", "fudge"]
 
 @client.event
 async def on_ready():
@@ -16,17 +17,37 @@ async def remove(ctx, amount=0):
     channel = client.get_channel(810409079481696299)
     await channel.send(str(amount) + " messages were removed.")
 
+#adds a word to the blacklist
+@client.command(pass_context=True)
+async def addWord(ctx, word):
+    blacklist.append(word)
+    await ctx.author.send(word + " has been added to the filter.")
+    await ctx.message.delete()
+
+#removes a word from the blacklist
+@client.command(pass_context=True)
+async def delWord(ctx, word):
+    blacklist.remove(word)
+    await ctx.author.send(word + " has been removed from the filter.")
+    await ctx.message.delete()
+
+#sends blacklisted words to user
+@client.command(pass_context=True)
+async def wordList(ctx):
+    await ctx.author.send("The blacklisted words are: " + str(blacklist))
+
 #check to see if one of the blacklisted words is found in a user's message, and if so it will be removed
 @client.event
 async def on_message(message):
-    blacklist = ["frick", "dang", "darn", "damn", "heck", "crap", "shizzle", "fudge"]
     channel = client.get_channel(810409079481696299)
     author_id = message.author.id
-    for word in blacklist:
-        if(word in message.content):
-            await message.delete()
-            await channel.send(getRandomMessage(author_id))
-    await client.process_commands(message)
+    delKeyword = "!delWord"
+    if((author_id != 810408611573661726)):
+        for word in blacklist:
+            if((word in message.content) and (delKeyword not in message.content)):
+                await message.delete()
+                await channel.send(getRandomMessage(author_id))
+        await client.process_commands(message)
 
 #this function allows the bot to send a snarky reply after a user swears
 def getRandomMessage(user_id):
@@ -43,4 +64,4 @@ def getRandomMessage(user_id):
     reply = reply.format(user_id)
     return reply
 
-client.run("SECRET TOKEN")
+client.run("ODEwNDA4NjExNTczNjYxNzI2.YCjN1Q.HE7qkKrlww6TpIy5LSUsJhiH_eE")
